Modified "Spirits of the Dead" to be less spooky
Although this may be an ongoing effort, these changes are complete
enough to allow the entire poem to still make sense. If this were
code, we would ensure that the commit continues to compile.

I updated the poem with my text editor, then just ran:
  git commit -a
The "-a" option tells Git to autoatically stage any files that
were modified or deleted, This option would not be sufficient
if I added new files too.