More changes to make "Sprits of the Dead" less spooky, and legal
These changes aesthetically depend upon the previous commit, but are
logically distinct, so I'm keeping this as a seperate commit.

Also added a copyright file to clarify the license of the
now-derivative work. Since this is a new file, simply running:
  git commit -a
will not include it in the final commit. I have to run:
  git add -A
then
  git commit