Renamed "Sprits of the Dead" to "Spirits of the Departed"
Git stores the rename / move of a file as a deletion of the old name,
then a creation of the new name. However, Git's tools are smart enough
to realize that this is semantically a rename operation if the content
of the files are identical or even similar.

I have to use:
  git add -A
to stage all of the changes (the deleted file + the new file). Once
staged, Git will start to see change as a rename.

Then I commit normally with
  git commit
Since this is a minor change, I am breaking the rules and committing
it directly to master. Try and stop me!